(feat) Add ability to disable and set default values for answers in specificQuestions
